+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Renamed host and guest respectively to management and workload.
+
 ## [1.2.0] - 2020-12-18
 
 ### Removed